float changes

+++ calculate.java
@@ -1,9 +1,9 @@
 public class calculate{
 	public static void main(String[] arg){
 		System.out.println("Calculate...");
-		double first = Double.valueOf(arg[0]);
-		double second = Double.valueOf(arg[1]);
-		double sum = first + second;
+		float first = Float.valueOf(arg[0]);
+		float second = Float.valueOf(arg[1]);
+		float sum = first + second;
 		System.out.println("Summ " + sum);		
 	}
 }
